refactor(server): apply A+ rubric — drop theatre, fix dep hygiene
Technical correctness:
- liquibase-hibernate7 5.0.2-rc1 → 5.0.3 (stable GA)
- Drop dead spring-modulith-bom (no Modulith API used after theatre rip-out)
- Add okhttp-bom for version coherence; drop redundant per-artifact versions
- Add spring-boot-starter-aspectj (Boot 4 rename of starter-aop); @Retryable
  was working via transitive aspectj only — now explicit
- Broaden Jackson FormatMapper bridge to ObjectMapper (only ObjectMapper API
  used; broader bean type makes it less brittle against Boot 4 patch versions)
- Add HALT preconditions to the 3 addNotNullConstraint changesets so deploys
  surface NULL-row issues with a clear message before ALTER TABLE locks the table

Cleanup:
- Delete 2 of 3 Jackson3FormatMapperTest methods (round-trip + null short-circuit
  were testing Jackson, not the 50-line bridge)
- Split SecurityFilterChainArchitectureTest into 2 top-level classes; collapse
  custom 22-line ArchCondition to one-line beDeclaredIn (-90 lines)
- Strip migration-narration comments (PiResultParser tokenStreamFactory rename,
  PracticeDetectionResultParser duplicate "lenient mapper", GitHubGraphQlConfig
  mixin restatement, SecurityConfig CSRF-disabled redundancy, pom.xml
  testcontainers-renamed and okhttp-jvm verbose comments, ChatMessage parts
  5-line Javadoc → 1, TaskEnvelopeWriter test-name reference, PracticeFindingDetailDTO.from
  signature-restatement Javadoc)
- Deduplicate enum-config comment (kept in AchievementRegistry where the code
  lives, removed from application.yml)
- Drop dead jackson.enum YAML config (false is the Jackson default)
- Simplify PracticeFindingDetailDTO.toMap triple-null-guard to single null check
- Add TODO in Jackson3FormatMapper for Hibernate 7.3 native replacement

Net: -140 lines.
